Fold BoutiquePack into Pack route
