+++ packages/apps/browser-app/src/translations/compiled/en.json
@@ -243,16 +243,6 @@
       "value": "Boutique"
     }
   ],
-  "BoutiquePack.tsx_AHlDyN": [
-    {
-      "type": 0,
-      "value": "Boutique » "
-    },
-    {
-      "type": 1,
-      "value": "packId"
-    }
-  ],
   "BrowserApp.tsx_w1vfWS": [
     {
       "type": 0,
@@ -2630,6 +2620,16 @@
     {
       "type": 0,
       "value": " to null"
+    }
+  ],
+  "Pack.tsx_AHlDyN": [
+    {
+      "type": 0,
+      "value": "Boutique » "
+    },
+    {
+      "type": 1,
+      "value": "packId"
     }
   ],
   "PackCard.tsx_EgRJme": [
